Major refactor of ball behavior
The 2048 specific features of ball behavior have been moved to the game controller, so the ball behavior just handles the collision and merging, and the game controller can set the custom behavior. This way, we can override almost everything with a new game mode, making implementation of new variations super easy.