Fixed fibonacci double merge bug
When two balls hit and decided to merge, the first would merge 1 + 1 into 2, but the other would be deleted at end of frame. Before that happened, the ball to be destroyed would see itself as 1 and the partner as 2, and merge into 3. We get around this by setting the value of the ball to be destroyed to -1 in all gamemodes, so it won't double detect 1 1 2 like that. Also need a try/catch to catch the -1 not being a fib number.